compilation probs
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7538 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/playing.cxx
@@ -728,7 +728,7 @@ static void		doKeyPlaying(const BzfKeyEvent& key, bool pressed)
   int d = temp.size() - 1;
   while (!isspace(temp[d]) && d >= 0) //get the pos of the start of the seed callsign
     --d;
-  while (d != temp.size()) { //now just load the seed callsign, which finishes the string
+  while (d != (int)temp.size()) { //now just load the seed callsign, which finishes the string
     name += temp[d];
     ++d;
   }
